[ISSUE #797]🎨Extract NamespaceUtil from the protocol🔥
